Shape, ShapeContainer, and Shape classes moved to seperate files.
